busca-app/backend/requirements.txt
Scrapy
Playwright
scrapy-playwright
playwright-stealth
django
djangorestframework
django-cors-headers
mysqlclient
undetected-chromedriver
Pillow
pytest
coverage
pytest-cov
pytest-asyncio